requirements.txt
cloudpickle==1.6.0
croniter==1.0.12
mlflow==1.14.1
six==1.14.0
PyYAML==5.4.1
pyspark==3.1.2
pyarrow==1.0.1
pandas==1.0.1
boto3==1.17.62
hyperopt==0.2.5
s3fs==0.4.2
psycopg2-binary==2.8.4
hvac==0.10.10
numpy==1.18.1
pendulum==1.4.4
sqlalchemy==1.3.23
flake8==3.9.1
slackclient==2.1.0
click==7.1.2
click-shell==2.0
databricks-cli==0.14.3
google-auth==1.21.2
google-api-python-client==1.12.1
alembic==1.4.1
deprecation==2.1.0
pydantic==1.7.4
mypy==0.812
sqlalchemy-stubs==0.4
requests==2.24.0
pyspark-stubs==2.3.0.post2
delta-spark==1.0.0
black==21.7b0
pytest==6.2.4
coverage==5.5
ruamel.yaml==0.17.9
pytest-dependency==0.5.1
